Add initial config set up method
There could be a few NeoFS network parameters. It
is inefficient to set them one by one, especally
in network initializing. To solve this issue there is
a new method "InitConfig" in the neofs contract.

This method works the same way as "Deploy" method:
it can be invoked only once. This methods gets any
number of key value pairs in the arguments.