nns: Allow to register pre-defined TLDs on deployment stage
In order to register TLD, committee multi-signature must be gathered.
It is not always easy. For example, on "fresh" network, committee
members may not have a notary role, so they will not be able to use
Notary service to collect a signature. At the same time, Notary service
is the only convenient way to collect a multi-signature in a distributed
mode. To simplify the solution of the described task, it would be
convenient to be able to register pre-known TLDs at the NNS contract
deployment stage (which is performed by the committee).

Support list of (name, e-mail) pairs describing pre-defined TLDs as
optional deployment parameters (passed into `_deploy` callback). Also
make `migration.NewContract` to pre-register 'neofs' TLD.

Refs #334.

Signed-off-by: Leonard Lyubich <leonard@morphbits.io>