nns: Add dedicated method for TLD registration
Top-level domains are controlled by the committee. Previously, NNS
contract provided `register` method that accepted fixed domain owner.
After recent changes, TLDs were forbidden to be treated as regular
tokens in terms of the NNS contract. According to this, `register`
method is no longer well-suited for TLD registration, so, it's worth to
provide dedicated method for this purpose.

Add `registerTLD` method with signature similar to the `register` one
but w/o owner parameter.

Closes #334.

Signed-off-by: Leonard Lyubich <leonard@morphbits.io>