deploy: transfer GAS in the first place if notaries are set
It is possible to start blockchain with required roles already set. In this
case, there is no need to wait for NNS to be deployed for GAS transfers, since
notary service (requires Notary roles to be set correctly) is ready. It allows
all multi-signed operations to be applied at an earlier stage. Closes #409.

Signed-off-by: Pavel Karpy <carpawell@nspcc.ru>