deploy: copy transaction expected to be changed
The whole `makeUnsignedDesignateCommitteeNotaryTx` is designed to be called many
times if something goes wrong. But currently neo-go's actor shares signers with
TXs it produces, so the actor is broken immediately after this func is called.
Can be seen only if transaction-to-sign is expired, in practice happened when
deployment nodes were started with a huge delay.

Signed-off-by: Pavel Karpy <carpawell@nspcc.ru>