netmap: add new node format, fix #298, fix #297
This format is added in parallel to existing one because we can't migrate in
the contract easily, it'd require complete protobuf parser in VM. Instead we
provide a secondary map here that works similar to the old one (except it
doesn't have any listing or duplication issues). The expection is that nodes
will send _both_ transactions starting from some time, we'll monitor them and
transition to using new map only in subsequent IR/SN release. Then contract
could be updated to drop the old data.

Signed-off-by: Roman Khimov <roman@nspcc.ru>