container: Do not persist container and eACL credentials
They are ony passed from creators to IR in notary requests at creation
stage. Once IR approves the container/eACL, they are persisted and
trusted. So, it makes no sense to drag credentials around.

This makes Container contract to store binaries only. Credentials are
dropped and storage is reformatted on update as usual.

Refs #481.

Signed-off-by: Leonard Lyubich <leonard@morphbits.io>