workflows: build with 1.24 only
The thing we're building here is contracts. They're using NeoGo compiler, so
outer Go version is not very much relevant (although can theoretically be in
which case we'd like to be consistent since we have binaries stored in the
repo). It's useful for tests, but tests do use two versions anyway, so it's
safe to drop one run here and simplify things.

Signed-off-by: Roman Khimov <roman@nspcc.ru>