container: Store containers as structures
This declares struct for containers similar to
71fc8a9e61dec7700690f2e6cad04326b80e1e13. Now containers are stored and
available for reading in two formats - structured and protobuf. The
former is available via new `getInfo` call. The latter is retained
for safety. It will be removed in the future.

New creation method `createV2` accepts structured container and replaces
all other methods.

Structured items for already existing containers is done iteratively by
Alphabet calls to the new `addStructs` method.

Refs #449.

Signed-off-by: Leonard Lyubich <leonard@morphbits.io>