nns: add new `Neo` record type
Add a new record type that can store large amounts of data, with efficient
existence checking, as it is stored in a key. Due to the new data storage type
in NNS, provide new iterators `ResolveNeoIterator` and `GetNeoRecordsIterator`,
as the element limit in the context may be exceeded for array. Also add new
funcs `AddNeoRecord` to add new records and `DeleteNeoRecord` to delete single
address.

Signed-off-by: Andrey Butusov <andrey@nspcc.io>